Custom Tagger
In order to have no conflicts between libraries that would like
to use either a special transform or any other tagger method
the logic has been refactored to grant a single Tagger per each
customization that might be needed.

The goal is to both fix #60 and improve heresy logic and reliability.